CRUD sum up

+++ app/src/Controller/CategoryController.php
@@ -96,8 +96,8 @@ class CategoryController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $category->setCreatedAt(new \DateTime());
-            $category->setUpdatedAt(new \DateTime());
+//            $category->setCreatedAt(new \DateTime());
+//            $category->setUpdatedAt(new \DateTime());
             $categoryRepository->save($category);
 
             $this->addFlash('success', 'message_created_successfully');
@@ -172,8 +172,14 @@ class CategoryController extends AbstractController
      *     name="category_delete",
      * )
      */
-    public function delete(Request $request, Category $category, CategoryRepository $categoryRepository): Response
+    public function delete(Request $request, Category $category, CategoryRepository $repository): Response
     {
+        if ($category->getTasks()->count()) {
+            $this->addFlash('warning', 'message_category_contains_tasks');
+
+            return $this->redirectToRoute('category_index');
+        }
+
         $form = $this->createForm(FormType::class, $category, ['method' => 'DELETE']);
         $form->handleRequest($request);
 
@@ -182,8 +188,8 @@ class CategoryController extends AbstractController
         }
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $categoryRepository->delete($category);
-            $this->addFlash('success', 'message.deleted_successfully');
+            $repository->delete($category);
+            $this->addFlash('success', 'message_deleted_successfully');
 
             return $this->redirectToRoute('category_index');
         }
